LayoutTests/http/tests/w3c/html/semantics/embedded-content/media-elements/synchronising-multiple-media-elements/media-controllers/task-source-expected.txt


FAIL MediaController events task source assert_array_equals: property 1, expected object "[object MediaController]" but got object "[object MediaController]"
